+++ sdk/resourcemanager/resourcehealth/armresourcehealth/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History
 
+## 1.0.0 (2022-05-18)
+### Breaking Changes
+
+- Function `OperationListResult.MarshalJSON` has been removed
+- Function `AvailabilityStatusListResult.MarshalJSON` has been removed
+- Function `ImpactedResourceStatusProperties.MarshalJSON` has been removed
+- Function `AvailabilityStatusPropertiesRecentlyResolved.MarshalJSON` has been removed
+- Function `ServiceImpactingEvent.MarshalJSON` has been removed
+- Function `StatusBanner.MarshalJSON` has been removed
+- Function `AvailabilityStatusProperties.MarshalJSON` has been removed
+
+
 ## 0.6.0 (2022-04-18)
 ### Breaking Changes